+++ src/test/java/org/folio/api/consortiumsearch/ConsortiumSearchLocationsIT.java
@@ -129,7 +129,8 @@ class ConsortiumSearchLocationsIT extends BaseConsortiumIntegrationTest {
       .flatMap(location -> Stream.of(
         kafkaResourceEvent(CENTRAL_TENANT_ID, CREATE, location, null),
         kafkaResourceEvent(MEMBER_TENANT_ID, CREATE, location, null)))
-      .forEach(event -> kafkaTemplate.send(inventoryLocationTopic(event.getTenant()), event));
+      .forEach(event -> kafkaTemplate.send(inventoryLocationTopic(event.getTenant()),
+        objectMapper.writeValueAsString(event)));
 
     await().atMost(ONE_MINUTE).pollInterval(ONE_SECOND).untilAsserted(() -> {
       var totalHits = countIndexDocument(LOCATION, CENTRAL_TENANT_ID);