+++ src/test/java/org/folio/support/base/BaseIntegrationTest.java
@@ -49,6 +49,7 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.function.Consumer;
 import lombok.SneakyThrows;
+import org.apache.commons.lang3.tuple.Pair;
 import org.awaitility.Awaitility;
 import org.awaitility.core.ThrowingRunnable;
 import org.folio.search.SearchApplication;
@@ -82,15 +83,21 @@ import org.opensearch.index.reindex.DeleteByQueryRequest;
 import org.opensearch.search.SearchHit;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.webmvc.test.autoconfigure.AutoConfigureMockMvc;
 import org.springframework.cache.CacheManager;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Import;
+import org.springframework.context.annotation.Primary;
 import org.springframework.http.HttpHeaders;
 import org.springframework.kafka.core.KafkaTemplate;
+import org.springframework.kafka.core.ProducerFactory;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.ResultMatcher;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
+import tools.jackson.databind.ObjectMapper;
 
 @EnableOkapi
 @EnableKafka
@@ -99,6 +106,8 @@ import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilde
 @AutoConfigureMockMvc
 @DirtiesContext(classMode = AFTER_CLASS)
 @SpringBootTest(classes = SearchApplication.class)
+@Import({
+  BaseIntegrationTest.KafkaTemplateTestConfiguration.class})
 public abstract class BaseIntegrationTest {
 
   protected static final String[] COLLECTION_IGNORING_FIELDS = {"items.id"};
@@ -106,7 +115,8 @@ public abstract class BaseIntegrationTest {
 
   protected static MockMvc mockMvc;
   protected static InventoryApi inventoryApi;
-  protected static KafkaTemplate<String, ResourceEvent> kafkaTemplate;
+  protected static KafkaTemplate<String, String> kafkaTemplate;
+  protected static ObjectMapper objectMapper;
   protected static OkapiConfiguration okapi;
   protected static RestHighLevelClient elasticClient;
   protected static CacheManager cacheManager;
@@ -358,7 +368,8 @@ public abstract class BaseIntegrationTest {
   @SneakyThrows
   protected static void setUpTenant(int expectedCount, Authority... authorities) {
     setUpTenant(TENANT_ID, authoritySearchPath(), () -> { }, asList(authorities), expectedCount, emptyList(),
-      rec -> kafkaTemplate.send(inventoryAuthorityTopic(), rec.getId(), resourceEvent(null, AUTHORITY, rec)));
+      rec -> kafkaTemplate.send(inventoryAuthorityTopic(), rec.getId(),
+        objectMapper.writeValueAsString(resourceEvent(null, AUTHORITY, rec))));
   }
 
   @SafeVarargs
@@ -388,7 +399,8 @@ public abstract class BaseIntegrationTest {
 
       if (type.equals(Authority.class)) {
         saveRecords(tenant, authoritySearchPath(), testRecords, expectedCount,
-          rec -> kafkaTemplate.send(inventoryAuthorityTopic(tenant), resourceEvent(null, AUTHORITY, rec)));
+          rec -> kafkaTemplate.send(inventoryAuthorityTopic(tenant),
+            objectMapper.writeValueAsString(resourceEvent(null, AUTHORITY, rec))));
       }
     }
   }
@@ -423,28 +435,10 @@ public abstract class BaseIntegrationTest {
   @SneakyThrows
   protected static void setUpTenant(Class<?> type, String tenant, Runnable postInitAction, Integer expectedCount,
                                     List<ResultMatcher> matchers, Map<String, Object>... records) {
-    String searchPath;
-    Consumer<Object> consumer;
 
-    if (type.equals(Instance.class)) {
-      searchPath = instanceSearchPath();
-      consumer = instance -> inventoryApi.createInstance(tenant, (Map<String, Object>) instance);
-    } else if (type.equals(Authority.class)) {
-      searchPath = authoritySearchPath();
-      consumer = authority -> kafkaTemplate.send(inventoryAuthorityTopic(tenant), event(authority, AUTHORITY, tenant));
-    } else if (type.equals(LinkedDataInstance.class)) {
-      searchPath = linkedDataInstanceSearchPath();
-      consumer = ldInstance -> kafkaTemplate.send(linkedDataInstanceTopic(tenant),
-        event(ldInstance, LINKED_DATA_INSTANCE, tenant));
-    } else if (type.equals(LinkedDataWork.class)) {
-      searchPath = linkedDataWorkSearchPath();
-      consumer = ldWork -> kafkaTemplate.send(linkedDataWorkTopic(tenant), event(ldWork, LINKED_DATA_WORK, tenant));
-    } else if (type.equals(LinkedDataHub.class)) {
-      searchPath = linkedDataHubSearchPath();
-      consumer = ldHub -> kafkaTemplate.send(linkedDataHubTopic(tenant), event(ldHub, LINKED_DATA_HUB, tenant));
-    } else {
-      throw new IllegalArgumentException("Unsupported type: " + type.getName());
-    }
+    var eventConsumer = getEventConsumer(type, tenant);
+    var searchPath = eventConsumer.getLeft();
+    var consumer = eventConsumer.getRight();
 
     setUpTenant(tenant, searchPath, postInitAction, asList(records), expectedCount, matchers, consumer);
   }
@@ -527,16 +521,39 @@ public abstract class BaseIntegrationTest {
     return resourceEvent(tenant, resourceType, CREATE, object);
   }
 
+  private static Pair<String, Consumer<Object>> getEventConsumer(Class<?> type, String tenant) {
+    if (type.equals(Instance.class)) {
+      return Pair.of(instanceSearchPath(), instance ->
+        inventoryApi.createInstance(tenant, (Map<String, Object>) instance));
+    } else if (type.equals(Authority.class)) {
+      return Pair.of(authoritySearchPath(), authority -> kafkaTemplate.send(inventoryAuthorityTopic(tenant),
+        objectMapper.writeValueAsString(event(authority, AUTHORITY, tenant))));
+    } else if (type.equals(LinkedDataInstance.class)) {
+      return Pair.of(linkedDataInstanceSearchPath(), ldInstance -> kafkaTemplate.send(linkedDataInstanceTopic(tenant),
+        objectMapper.writeValueAsString(event(ldInstance, LINKED_DATA_INSTANCE, tenant))));
+    } else if (type.equals(LinkedDataWork.class)) {
+      return Pair.of(linkedDataWorkSearchPath(), ldWork -> kafkaTemplate.send(linkedDataWorkTopic(tenant),
+        objectMapper.writeValueAsString(event(ldWork, LINKED_DATA_WORK, tenant))));
+    } else if (type.equals(LinkedDataHub.class)) {
+      return Pair.of(linkedDataHubSearchPath(), ldHub -> kafkaTemplate.send(linkedDataHubTopic(tenant),
+        objectMapper.writeValueAsString(event(ldHub, LINKED_DATA_HUB, tenant))));
+    } else {
+      throw new IllegalArgumentException("Unsupported type: " + type.getName());
+    }
+  }
+
   @BeforeAll
   static void setUpDefaultTenant(
     @Autowired MockMvc mockMvc,
-    @Autowired KafkaTemplate<String, ResourceEvent> kafkaTemplate,
+    @Autowired KafkaTemplate<String, String> kafkaTemplate,
+    @Autowired ObjectMapper objectMapper,
     @Autowired RestHighLevelClient restHighLevelClient,
     @Autowired CacheManager cacheManager) {
     setEnvProperty("folio-test");
     BaseIntegrationTest.mockMvc = mockMvc;
     BaseIntegrationTest.kafkaTemplate = kafkaTemplate;
-    BaseIntegrationTest.inventoryApi = new InventoryApi(kafkaTemplate);
+    BaseIntegrationTest.objectMapper = objectMapper;
+    BaseIntegrationTest.inventoryApi = new InventoryApi(kafkaTemplate, objectMapper);
     BaseIntegrationTest.elasticClient = restHighLevelClient;
     BaseIntegrationTest.cacheManager = cacheManager;
   }
@@ -551,5 +568,16 @@ public abstract class BaseIntegrationTest {
     removeEnvProperty();
   }
 
-  public record TestData(Class<?> type, List<Map<String, Object>> testRecords, int expectedCount) { }
+  public record TestData(Class<?> type, List<Map<String, Object>> testRecords, int expectedCount) {
+  }
+
+  @TestConfiguration
+  public static class KafkaTemplateTestConfiguration {
+
+    @Bean
+    @Primary
+    public KafkaTemplate<String, String> kafkaStringTemplate(ProducerFactory<String, String> producerFactory) {
+      return new KafkaTemplate<>(producerFactory);
+    }
+  }
 }